intrastat_base: Remove intrastat.common class
