+++ intrastat_base/models/__init__.py
@@ -1,5 +1,5 @@
 from . import product_template
 from . import res_company
-from . import intrastat_common
 from . import account_fiscal_position
+from . import account_fiscal_position_template
 from . import account_move
