+++ intrastat_base/models/res_company.py
@@ -2,8 +2,17 @@
 # @author: <alexis.delattre@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from odoo import _, api, fields, models
-from odoo.exceptions import ValidationError
+import logging
+from io import BytesIO
+from sys import exc_info
+from traceback import format_exception
+
+from lxml import etree
+
+from odoo import _, api, fields, models, tools
+from odoo.exceptions import UserError, ValidationError
+
+logger = logging.getLogger(__name__)
 
 
 class ResCompany(models.Model):
@@ -40,3 +49,24 @@ class ResCompany(models.Model):
                     raise ValidationError(
                         _("Missing e-mail address on user '%s'.") % (user.name)
                     )
+
+    @api.model
+    def _intrastat_check_xml_schema(self, xml_bytes, xsd_file):
+        """Validate the XML file against the XSD"""
+        xsd_etree_obj = etree.parse(tools.file_open(xsd_file, mode="rb"))
+        official_schema = etree.XMLSchema(xsd_etree_obj)
+        try:
+            t = etree.parse(BytesIO(xml_bytes))
+            official_schema.assertValid(t)
+        except (etree.XMLSchemaParseError, etree.DocumentInvalid) as e:
+            logger.warning("The XML file is invalid against the XML Schema Definition")
+            logger.warning(xml_bytes)
+            logger.warning(e)
+            usererror = "{}\n\n{}".format(e.__class__.__name__, str(e))
+            raise UserError(usererror)
+        except Exception:
+            error = _("Unknown Error")
+            tb = "".join(format_exception(*exc_info()))
+            error += "\n%s" % tb
+            logger.warning(error)
+            raise UserError(error)
