--- intrastat_base/tests/data/mail_template.xml
@@ -1,59 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!--
-  Copyright 2013-2020 Akretion France (http://www.akretion.com/)
-  @author Alexis de Lattre <alexis.delattre@akretion.com>
-  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
--->
-<odoo>
-        <record
-        id="base_intrastat_product_reminder_email_template"
-        model="mail.template"
-    >
-            <field name="name">Intrastat Product Reminder</field>
-            <field
-            name="model_id"
-            ref="intrastat_base.model_intrastat_declaration_test"
-        />
-            <field name="auto_delete" eval="False" />
-            <field
-            name="email_from"
-        >${object.company_id.email or 'odoo@example.com'}</field>
-            <field name="email_to">${object.company_id.intrastat_email_list}</field>
-            <field
-            name="subject"
-        >${object.type} DEB ${object.year_month} for ${object.company_id.name}</field>
-            <field name="body_html">
-                <![CDATA[
-<div style="font-family: 'Lucica Grande', Ubuntu, Arial, Verdana, sans-serif; font-size: 12px; color: rgb(34, 34, 34); background-color: #FFF; ">
-
-<p>I would like to remind you that we are approaching the deadline for the DEB for month ${object.year_month}.</p>
-
-<p>As there were no ${object.type} DEB for that month in Odoo, a draft DEB has been generated automatically by Odoo.</p>
-
-% if ctx.get('exception'):
-<p>When trying to generate the lines of the ${object.declaration_type} DEB, the following error was encountered:</p>
-
-<p>${ctx.get('error_msg')}</p>
-
-<p>You should solve this error, then go to the menu "Invoicing > Reporting > Intrastat > DEB", open the ${object.declaration_type} declaration for month ${object.year_month} and click on the button "Generate lines from invoices".</p>
-
-% else:
-% if object.num_lines and object.num_lines > 0:
-<p>This draft ${object.type} DEB contains ${object.num_decl_lines} ${object.num_decl_lines == 1 and 'line' or 'lines'}.</p>
-% else:
-<p>This draft ${object.type} DEB generated automatically by Odoo doesn't contain any line.</p>
-% endif
-
-<p>Go and check this declaration in Odoo in the menu "Invoicing > Reporting > Intrastat > DEB".</p>
-
-% endif
-
-<p>
---
-Automatic e-mail sent by Odoo.
-</p>
-</div>
-]]>
-            </field>
-        </record>
-</odoo>
