+++ intrastat_product/__init__.py
@@ -1,2 +1,3 @@
 from . import models
 from . import report
+from . import wizards
