+++ intrastat_product/tests/test_intrastat_product.py
@@ -4,7 +4,7 @@ from psycopg2 import IntegrityError
 
 from odoo.tests.common import SavepointCase
 from odoo.tools import mute_logger
-
+from odoo.exceptions import UserError, ValidationError
 from .common import IntrastatProductCommon
 
 
@@ -38,6 +38,28 @@ class TestIntrastatProduct(IntrastatProductCommon):
         self._create_declaration(vals)
         decl_copy = self.declaration.copy()
         self.assertEqual(self.declaration.revision + 1, decl_copy.revision)
+
+    def test_declaration_no_country(self):
+        self.demo_company.country_id = False
+        with self.assertRaises(ValidationError):
+            self._create_declaration()
+            self.declaration.flush()
+
+    def test_declaration_no_vat(self):
+        self.demo_company.partner_id.vat = False
+        with self.assertRaises(UserError):
+            self._create_declaration()
+            self.declaration.flush()
+            self.declaration._check_generate_xml()
+
+    def test_declaration_state(self):
+        self._create_declaration()
+        self.declaration.unlink()
+
+        self._create_declaration()
+        self.declaration.state = "done"
+        with self.assertRaises(UserError):
+            self.declaration.unlink()
 
 
 class TestIntrastatProductCase(TestIntrastatProduct, SavepointCase):
