+++ intrastat_product/wizards/__init__.py
@@ -1,0 +1,1 @@
+from . import intrastat_result_view
