+++ intrastat_product/wizards/intrastat_result_view.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<!--
+  Copyright 2011-2021 Akretion France (http://www.akretion.com/)
+  Copyright 2015-2021 Noviat (http://www.noviat.com/)
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+-->
+<odoo>
+    <record id="intrastat_result_view_form" model="ir.ui.view">
+        <field name="name">intrastat.result_view_form</field>
+        <field name="model">intrastat.result.view</field>
+        <field name="arch" type="xml">
+            <form string="Intrastat Result View">
+                <group name="main">
+                    <field name="note" nolabel="1" />
+                </group>
+                <footer>
+                    <button string="Ok" class="btn-primary" special="cancel" />
+                </footer>
+            </form>
+        </field>
+    </record>
+</odoo>
